BASEscraping/requirements.txt
newspaper3k==0.2.8
pandas==2.1.1
nltk==3.8.1
openpyxl==3.1.2
pytz==2023.3.post1
gensim==4.3.2
urllib3==2.0.7
gspread==5.7.1
oauth2client==4.1.3
gspread_dataframe==3.3.1
requests==2.31.0